+++ sqlparser/lib/src/engine/sql_engine.dart
@@ -31,8 +31,6 @@ class SqlEngine {
     registerTable(sqliteSchema);
 
     registerTable(sqliteSequence);
-
-    registerTable(dbstat);
   }
 
   /// Obtain a [SchemaFromCreateTable] instance compatible with the