+++ drift_dev/CHANGELOG.md
@@ -7,6 +7,7 @@
   on the web.
 - Fix generated code missing question marks for nullable types in some instances.
 - Add (opt-in) `row_class_constructor_all_required` builder option.
+- Support the `dbstat` table in drift-file queries when enabling the `dbstat` module.
 
 ## 2.21.2